feat: mobile sidenav via attribute #13
We propose a generalized solution for managing state, content, and/or
structural changes based on variable screen resolutions via a
Boolean attribute, static, to dictate the rendering path.

The sidenav is now mobile by default and the static attribute removes
the accordion. The auro-accordion package is now a peer dependency.

For extended context please see:
https://github.com/AlaskaAirlines/auro-sidenav/discussions/19

BREAKING CHANGE: all auro-sidenav instances are now mobile until
specified with the static attribute.

Changes to be committed:
	modified:   apiExamples/basic.html
	modified:   apiExamples/custom.html
	modified:   apiExamples/full.html
	modified:   apiExamples/heading.html
	modified:   apiExamples/hyperlink.html
	modified:   apiExamples/isToggle.html
	modified:   apiExamples/sections.html
	modified:   apiExamples/selected.html
	modified:   demo/api.md
	modified:   demo/index.html
	modified:   demo/index.md
	modified:   docs/api.md
	modified:   docs/partials/index.md
	modified:   package-lock.json
	modified:   src/auro-sidenav.js
	modified:   src/auro-sidenavsection.js
	modified:   test/auro-sidenav.test.js
   modified:   apiExamples/collapsible.html
   modified:   README.md